requirements.txt
Flask==2.3.2
Werkzeug==2.3.6
groq==0.4.2
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
openai==1.25.0
bcrypt==4.1.1
email-validator==2.1.0
